ramadan-coding-challenge/QA-Chatbot/pyproject.toml
[project]
name = "qa-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.1",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
]
